refactor(agent): centralize context graph authority model
