+++ packages/agent/package.json
@@ -20,6 +20,12 @@
     },
     "./package.json": "./package.json",
     "./dist/internal/*": null,
+    "./dist/context-graph-authority.js": null,
+    "./dist/context-graph-authority.*": null,
+    "./dist/context-graph-agent-gate-authority.js": null,
+    "./dist/context-graph-agent-gate-authority.*": null,
+    "./dist/context-graph-access-policy.js": null,
+    "./dist/context-graph-access-policy.*": null,
     "./dist/rfc64/author-catalog-producer.js": "./dist/rfc64/author-catalog-producer.js",
     "./dist/rfc64/catalog-row-authorship.js": "./dist/rfc64/catalog-row-authorship.js",
     "./dist/rfc64/inventory-v1/candidate.js": "./dist/rfc64/inventory-v1/candidate.js",
