+++ packages/agent/scripts/test-package-root.mjs
@@ -60,22 +60,28 @@ if (
 ) {
   throw new Error('internal authority marker machinery leaked from the package root');
 }
-const internalPromoteModule = 'internal/promote/context-graph-agent-gate-authority';
-// One structural namespace rule protects current and future implementation files.
+const internalModules = [
+  'internal/promote/assertion-promote-precommit',
+  'context-graph-authority',
+  'context-graph-agent-gate-authority',
+  'context-graph-access-policy',
+];
 // Check runtime imports and generated declaration/map paths against the built package.
-for (const extension of ['js', 'd.ts', 'js.map', 'd.ts.map']) {
-  const specifier = `@origintrail-official/dkg-agent/dist/${internalPromoteModule}.${extension}`;
-  try {
-    await import(specifier);
-    throw new Error(`internal promote module unexpectedly resolved: ${specifier}`);
-  } catch (error) {
-    if (error?.code !== 'ERR_PACKAGE_PATH_NOT_EXPORTED') throw error;
-  }
-  try {
-    require.resolve(specifier);
-    throw new Error(`internal promote module unexpectedly resolved via require: ${specifier}`);
-  } catch (error) {
-    if (error?.code !== 'ERR_PACKAGE_PATH_NOT_EXPORTED') throw error;
+for (const internalModule of internalModules) {
+  for (const extension of ['js', 'd.ts', 'js.map', 'd.ts.map']) {
+    const specifier = `@origintrail-official/dkg-agent/dist/${internalModule}.${extension}`;
+    try {
+      await import(specifier);
+      throw new Error(`internal module unexpectedly resolved: ${specifier}`);
+    } catch (error) {
+      if (error?.code !== 'ERR_PACKAGE_PATH_NOT_EXPORTED') throw error;
+    }
+    try {
+      require.resolve(specifier);
+      throw new Error(`internal module unexpectedly resolved via require: ${specifier}`);
+    } catch (error) {
+      if (error?.code !== 'ERR_PACKAGE_PATH_NOT_EXPORTED') throw error;
+    }
   }
 }
 const legacySynchronizationError = new legacyCatalogSync.Rfc64CatalogSynchronizationErrorV1(
