+++ packages/agent/src/dkg-agent-crypto.ts
@@ -138,14 +138,16 @@ import {
   type ActivePublicContextGraphChainProof,
 } from './active-public-context-graph-chain-proof.js';
 import {
-  createContextGraphAuthorityError,
   resolveContextGraphAgentGateAuthorityDecision,
-  type ContextGraphAgentGateAuthority,
-} from './internal/promote/context-graph-agent-gate-authority.js';
+} from './context-graph-agent-gate-authority.js';
 import {
   resolveLiveOnChainAccessPolicyState as resolveLiveAccessPolicyState,
+} from './context-graph-access-policy.js';
+import {
+  createContextGraphAuthorityError,
+  type ContextGraphAgentGateAuthority,
   type LiveOnChainAccessPolicyState,
-} from './internal/promote/context-graph-access-policy-state.js';
+} from './context-graph-authority.js';
 
 import { ProfileManager } from './profile-manager.js';
 import { DiscoveryClient, type SkillSearchOptions, type DiscoveredAgent, type DiscoveredOffering } from './discovery.js';
