+++ packages/agent/src/registered-context-graph-participant-mutation.ts
@@ -2,7 +2,7 @@
 
 import type { ChainAdapter, TxResult } from '@origintrail-official/dkg-chain';
 import { ethers } from 'ethers';
-import type { RegisteredContextGraphAuthority } from './dkg-agent-cg-resolve.js';
+import type { RegisteredContextGraphAuthority } from './context-graph-authority.js';
 
 export type RegisteredParticipantMutationOperation = 'add' | 'remove';
